requirements.txt
accelerate==1.12.0
annotated-doc==0.0.4
anyio==4.11.0
backports.tarfile==1.2.0
black==25.9.0
build==1.3.0
CacheControl==0.14.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cleo==2.1.0
click==8.3.0
crashtest==0.4.1
cryptography==46.0.3
cuda-bindings==12.9.4
cuda-pathfinder==1.3.5
distlib==0.4.0
duckdb==1.4.4
dulwich==0.24.4
faiss-cpu==1.13.2
fastjsonschema==2.21.2
filelock==3.20.0
findpython==0.7.0
flatbuffers==25.12.19
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.7.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jeepney==0.9.0
Jinja2==3.1.6
joblib==1.5.3
keyring==25.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
mypy_extensions==1.1.0
networkx==3.4.2
ninja==1.13.0
nncf==3.0.0
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.590.48
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
onnx==1.20.1
onnxruntime-gpu==1.24.2
openvino==2026.0.0
openvino-telemetry==2025.2.0
openvino-tokenizers==2026.0.0.0
optimum==2.1.0
optimum-intel==1.27.0
optimum-onnx==0.1.0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pbs-installer==2025.10.14
pkginfo==1.12.1.2
platformdirs==4.5.0
poetry==2.2.1
poetry-core==2.2.1
polars==1.38.1
polars-runtime-32==1.38.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.1
pybind11==3.0.1
pycparser==2.23
pydot==3.0.4
Pygments==2.19.2
pyparsing==3.3.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.1
regex==2026.2.19
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
SecretStorage==3.4.0
sentence-transformers==5.2.3
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tabulate==0.9.0
tensorrt-cu12==10.15.1.29
threadpoolctl==3.6.0
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.10.0
tqdm==4.67.3
transformers==4.57.6
triton==3.6.0
trove-classifiers==2025.9.11.17
typer==0.24.1
typer-slim==0.24.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.5.0
virtualenv==20.35.3
zipp==3.23.0
zstandard==0.25.0
